Add github installation event handler
