+++ src/Clients/PullRequestQuantifier.GitHub.Client/Events/GitHubEventHost.cs
@@ -1,40 +1,35 @@
 ﻿namespace PullRequestQuantifier.GitHub.Client.Events
 {
     using System;
-    using System.Drawing;
-    using System.IO;
-    using System.Text;
+    using System.Collections.Generic;
+    using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.Extensions.Hosting;
-    using Octokit;
-    using PullRequestQuantifier.Abstractions.Core;
-    using PullRequestQuantifier.Abstractions.Git;
-    using PullRequestQuantifier.Client.Extensions;
-    using PullRequestQuantifier.Client.QuantifyClient;
-    using PullRequestQuantifier.GitHub.Client.GitHubClient;
+    using Newtonsoft.Json.Linq;
+    using PullRequestQuantifier.GitHub.Client.Models;
     using PullRequestQuantifier.GitHub.Client.Telemetry;
 
     public class GitHubEventHost : IHostedService
     {
         private readonly IEventBus eventBus;
         private readonly IAppTelemetry appTelemetry;
-        private readonly IGitHubClientAdapterFactory gitHubClientAdapterFactory;
+        private readonly IDictionary<GitHubEventActions, IGitHubEventHandler> gitHubEventHandlers;
 
         public GitHubEventHost(
             IEventBus eventBus,
             IAppTelemetry appTelemetry,
-            IGitHubClientAdapterFactory gitHubClientAdapterFactory)
+            IEnumerable<IGitHubEventHandler> gitHubEventHandlers)
         {
             this.eventBus = eventBus;
             this.appTelemetry = appTelemetry;
-            this.gitHubClientAdapterFactory = gitHubClientAdapterFactory;
+            this.gitHubEventHandlers = gitHubEventHandlers.ToDictionary(h => h.EventType);
         }
 
         public async Task StartAsync(CancellationToken cancellationToken)
         {
             await eventBus.SubscribeAsync(
-                QuantifyPullRequest,
+                HandleGitHubEvent,
                 ErrorHandler,
                 cancellationToken);
         }
@@ -44,126 +39,32 @@
             return Task.CompletedTask;
         }
 
-        private async Task QuantifyPullRequest(string pullRequestEvent)
+        private async Task HandleGitHubEvent(string gitHubEvent)
         {
-            var payload =
-                new Octokit.Internal.SimpleJsonSerializer().Deserialize<PullRequestEventPayload>(pullRequestEvent);
-
-            var gitHubClientAdapter =
-                await gitHubClientAdapterFactory.GetGitHubClientAdapterAsync(payload.Installation.Id);
-
-            // get pull request
-            var pullRequest = await gitHubClientAdapter.GetPullRequestAsync(
-                payload.Repository.Id,
-                payload.PullRequest.Number);
-
-            // get pull request files
-            var pullRequestFiles = await gitHubClientAdapter.GetPullRequestFilesAsync(
-                payload.Repository.Id,
-                payload.PullRequest.Number);
-
-            // convert to quantifier input
-            var quantifierInput = new QuantifierInput();
-            foreach (var pullRequestFile in pullRequestFiles)
+            var eventJtoken = JToken.Parse(gitHubEvent);
+            if (!Enum.TryParse(
+                eventJtoken["eventType"].ToString(),
+                true,
+                out GitHubEventActions parsedEvent))
             {
-                if (pullRequestFile.Patch == null)
-                {
-                    continue;
-                }
-
-                var changeType = GitChangeType.Modified;
-                switch (pullRequestFile.Status)
-                {
-                    case "modified":
-                        break;
-                    case "added":
-                        changeType = GitChangeType.Added;
-                        break;
-                    case "deleted":
-                        changeType = GitChangeType.Deleted;
-                        break;
-                }
-
-                var fileExtension = !string.IsNullOrWhiteSpace(pullRequestFile.FileName)
-                    ? new FileInfo(pullRequestFile.FileName).Extension
-                    : string.Empty;
-                var gitFilePatch = new GitFilePatch(
-                    pullRequestFile.FileName,
-                    fileExtension)
-                {
-                    ChangeType = changeType,
-                    AbsoluteLinesAdded = pullRequestFile.Additions,
-                    AbsoluteLinesDeleted = pullRequestFile.Deletions,
-                    DiffContent = pullRequestFile.Patch,
-                };
-                quantifierInput.Changes.Add(gitFilePatch);
+                return;
             }
 
-            // get context if present
-            string context = null;
-            try
+            if (gitHubEventHandlers.TryGetValue(parsedEvent, out var selectedEventHandler))
             {
-                var rawContext = await gitHubClientAdapter.GetRawFileAsync(
-                    payload.Repository.Owner.Login,
-                    payload.Repository.Name,
-                    "/prquantifier.yaml");
-                context = Encoding.UTF8.GetString(rawContext);
+                await selectedEventHandler.HandleEvent(gitHubEvent);
             }
-            catch (NotFoundException)
+            else
             {
+                // this should never happen
+                // there must always be a handler for an accepted eventType
+                // TODO: log
             }
-            catch
-            {
-                // ignored
-            }
-
-            var quantifyClient = new QuantifyClient(context);
-            var quantifierClientResult = await quantifyClient.Compute(quantifierInput);
-
-            // create a new label in the repository if doesn't exist
-            try
-            {
-                var existingLabel = await gitHubClientAdapter.GetLabelAsync(
-                    payload.Repository.Id,
-                    quantifierClientResult.Label);
-            }
-            catch (NotFoundException)
-            {
-                // create new label
-                var color = Color.FromName(quantifierClientResult.Color);
-                await gitHubClientAdapter.CreateLabelAsync(
-                    payload.Repository.Id,
-                    new NewLabel(quantifierClientResult.Label, ConvertToHex(color)));
-            }
-
-            // apply label to pull request
-            await gitHubClientAdapter.ApplyLabelAsync(
-                payload.Repository.Id,
-                payload.PullRequest.Number,
-                new[] { quantifierClientResult.Label });
-
-            // create a comment on the issue
-            var defaultBranch = payload.Repository.DefaultBranch;
-            var quantifierContextLink = $"{payload.Repository.HtmlUrl}/blob/{defaultBranch}/prquantifier.yaml";
-            var comment = await quantifierClientResult.ToMarkdownCommentAsync(
-                payload.Repository.HtmlUrl,
-                quantifierContextLink,
-                payload.PullRequest.HtmlUrl,
-                payload.PullRequest.User.Login);
-            await gitHubClientAdapter.CreateIssueCommentAsync(
-                payload.Repository.Id,
-                payload.PullRequest.Number,
-                comment);
         }
 
         private Task ErrorHandler(Exception exception)
         {
             return Task.CompletedTask;
-        }
-
-        private string ConvertToHex(Color c)
-        {
-            return c.R.ToString("X2") + c.G.ToString("X2") + c.B.ToString("X2");
         }
     }
 }
