+++ src/Clients/PullRequestQuantifier.GitHub.Client/Events/IGitHubEventHandler.cs
@@ -1,0 +1,12 @@
+namespace PullRequestQuantifier.GitHub.Client.Events
+{
+    using System.Threading.Tasks;
+    using PullRequestQuantifier.GitHub.Client.Models;
+
+    public interface IGitHubEventHandler
+    {
+        GitHubEventActions EventType { get; }
+
+        Task HandleEvent(string gitHubEvent);
+    }
+}
