+++ src/Clients/PullRequestQuantifier.GitHub.Client/Models/GitHubEventActions.cs
@@ -12,7 +12,10 @@ namespace PullRequestQuantifier.GitHub.Client.Models
         Synchronize = 1 << 1,
         Reopened = 1 << 2,
         Created = 1 << 3,
+        Added = 1 << 4,
+        Deleted = 1 << 5,
+        Removed = 1 << 6,
         Pull_Request = Opened | Reopened | Synchronize,
-        Installation = Created
+        Installation = Created | Added | Deleted | Removed
     }
 }
