+++ src/Clients/PullRequestQuantifier.GitHub.Client/Registrar.cs
@@ -19,23 +19,30 @@ namespace PullRequestQuantifier.GitHub.Client
             serviceCollection.Configure<GitHubAppSettings>(configuration.GetSection(nameof(GitHubAppSettings)));
             serviceCollection.Configure<AzureServiceBusSettings>(
                 configuration.GetSection(nameof(AzureServiceBusSettings)));
-            serviceCollection.AddSingleton<IGitHubJwtFactory>(sp =>
-            {
-                // register a GitHubJwtFactory used to create tokens to access github for a particular org on behalf  of the  app
-                var gitHubAppSettings = sp.GetRequiredService<IOptions<GitHubAppSettings>>().Value;
-                ArgumentCheck.ParameterIsNotNull(gitHubAppSettings, nameof(gitHubAppSettings));
+            serviceCollection.AddSingleton<IGitHubJwtFactory>(
+                sp =>
+                {
+                    // register a GitHubJwtFactory used to create tokens to access github for a particular org on behalf  of the  app
+                    var gitHubAppSettings = sp.GetRequiredService<IOptions<GitHubAppSettings>>().Value;
+                    ArgumentCheck.ParameterIsNotNull(gitHubAppSettings, nameof(gitHubAppSettings));
 
-                // Use GitHubJwt library to create the GitHubApp Jwt Token using our private certificate PEM file
-                return new GitHubJwtFactory(
-                    new StringPrivateKeySource(gitHubAppSettings.PrivateKey),
-                    new GitHubJwtFactoryOptions
-                    {
-                        AppIntegrationId = int.Parse(gitHubAppSettings.Id), // The GitHub App Id
-                        ExpirationSeconds = 600 // 10 minutes is the maximum time allowed
-                    });
-            });
+                    // Use GitHubJwt library to create the GitHubApp Jwt Token using our private certificate PEM file
+                    return new GitHubJwtFactory(
+                        new StringPrivateKeySource(gitHubAppSettings.PrivateKey),
+                        new GitHubJwtFactoryOptions
+                        {
+                            AppIntegrationId = int.Parse(gitHubAppSettings.Id), // The GitHub App Id
+                            ExpirationSeconds = 600 // 10 minutes is the maximum time allowed
+                        });
+                });
             serviceCollection.AddSingleton<IGitHubClientAdapterFactory, GitHubClientAdapterFactory>();
             serviceCollection.TryAddSingleton<IEventBus, AzureServiceBus>();
+            serviceCollection.TryAddEnumerable(
+                new[]
+                {
+                    ServiceDescriptor.Singleton<IGitHubEventHandler, PullRequestEventHandler>(),
+                    ServiceDescriptor.Singleton<IGitHubEventHandler, InstallationEventHandler>()
+                });
             serviceCollection.AddHostedService<GitHubEventHost>();
 
             serviceCollection.AddApmForWebHost(configuration, typeof(Registrar).Namespace);
